Module Client.MassInsert

type command

val empty : command

val set : ?⁠ex:int -> ?⁠px:int -> ?⁠nx:bool -> ?⁠xx:bool -> string -> string -> command

val del : string list -> command

Delete a key; returns the number of keys removed.

val expire : string -> int -> command

val hset : string -> string -> string -> command

val hdel : string -> string -> command

Removes the specified fields from the hash stored at key. Specified fields that do not exist within this hash are ignored.

val hget : string -> string -> command

val hincrby : string -> string -> int -> command

val write : connection -> command list -> reply list IO.t

val incr : string -> command

val decr : string -> command